redirectpolicy: Move to pkg/loadbalancer/legacy
Move the old redirect policy implementation to pkg/loadbalancer/legacy
to clearly mark it as being on the way out.
